+++ CHANGELOG.md
@@ -15,6 +15,8 @@ created.
 ### Added
 - ([#142](https://github.com/microsoft/hi-ml/pull/142)) Adding AzureML progress bar and diagnostics for batch loading
 
+- ([#138](https://github.com/microsoft/hi-ml/pull/138)) Guidelines and profiling for whole slide images.
+
 ### Changed
 
 ### Fixed